+++ src/hci/gnomex/controller/GetRequestList.java
@@ -98,9 +98,9 @@ public class GetRequestList extends GNomExCommand implements Serializable {
         Integer idLab = (Integer) row[13];
         Integer idCoreFacility = (Integer) row[14];
         String corePrepInstructions = (String) row[15];
-        Integer numberOfSamples = (Integer) row[16];
+        Integer numberOfSamples = (int) (long) row[16];
         String adminNotes = (String) row[17];
-        Integer numberOfWorkItems = (Integer) row[18];
+        Integer numberOfWorkItems = (int) (long) row[18];
 
         String requestStatus = dh.getRequestStatus(codeRequestStatus);
         String labName = Lab.formatLabNameFirstLast(labFirstName, labLastName);
